chore(seed): split seedWorkoutPlans into safe standalone script
Wydzielenie logiki built-in planów do osobnego pliku, by można było
uruchamiać tylko seed planów bez kasowania userów i ćwiczeń produkcyjnych.

Zmiany:
- Nowy plik prisma/seedWorkoutPlans.ts — eksportuje funkcję seedWorkoutPlans,
  zero side-effectów na importu (poprzedni seed.ts wywoływał main() top-level,
  więc samo importowanie skasowałoby userów).
- Nowy entrypoint prisma/seed-plans.ts — uruchamia tylko seedWorkoutPlans()
  + prisma.$disconnect. Bezpieczny dla produkcji.
- prisma/seed.ts: usunięta lokalna definicja, dodany import z seedWorkoutPlans.js.
- package.json: nowy skrypt `seed:plans`.

Użycie:
  npm run seed       # pełen seed (kasuje userów — TYLKO LOKALNIE!)
  npm run seed:plans # bezpiecznie dla produkcji — dodaje 5 built-in planów FBW
                     # i 1 globalne ćwiczenie, nie rusza userów/treningów
